+++ example/lib/main.dart
@@ -74,8 +74,9 @@ class _MainPageState extends State<MainPage> {
               ElevatedButton(
                 onPressed: () async {
                   var path = await TBIBDownloader().downloadFile(
-                    url: 'http://212.183.159.230/50MB.zip',
-                    fileName: 'dummy.zip',
+                    url: 'https://speed.hetzner.de/100MB.bin',
+                    receiveBytesAsMB: true,
+                    fileName: 'dummy.bin',
                     directoryName: 'test',
                     onReceiveProgress: ({int? receivedBytes, int? totalBytes}) {
                       if (!context.mounted) {