+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0
+
+* Fix display file size unit in notification
+* Can receive bytes as MB
+
+## 0.0.2
+
+* solve images in pub dev
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* init.
+